+++ plugin_ide.ui/src/com/googlecode/goclipse/ui/editor/CompareViewerCreator.java
@@ -1,0 +1,27 @@
+/*******************************************************************************
+ * Copyright (c) 2015 Bruno Medeiros and other Contributors.
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v1.0
+ * which accompanies this distribution, and is available at
+ * http://www.eclipse.org/legal/epl-v10.html
+ *
+ * Contributors:
+ *     Bruno Medeiros - initial API and implementation
+ *******************************************************************************/
+package com.googlecode.goclipse.ui.editor;
+
+import org.eclipse.compare.CompareConfiguration;
+import org.eclipse.compare.IViewerCreator;
+import org.eclipse.jface.viewers.Viewer;
+import org.eclipse.swt.widgets.Composite;
+
+import melnorme.lang.ide.ui.editor.LangTextMergeViewer;
+
+public class CompareViewerCreator implements IViewerCreator {
+	
+	@Override
+	public Viewer createViewer(Composite parent, CompareConfiguration config) {
+		return new LangTextMergeViewer(parent, config);
+	}
+	
+}